backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.44.2
torch>=2.3.0
captum==0.7.0
python-dotenv==1.0.1
pydantic==2.9.2